\docType{data}
\name{clrFacadeTypeName}
\alias{clrFacadeTypeName}
\title{Full type name of the main facade to the interop code written in C#}
\format{chr "Rclr.ClrFacade"}
\usage{
  clrFacadeTypeName
}
\description{
  Full type name of the main facade to the interop code
  written in C#
}
\keyword{datasets}

