% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{clrGetConstructors}
\alias{clrGetConstructors}
\title{List the public constructors of a CLR Type}
\usage{
clrGetConstructors(type)
}
\arguments{
\item{type}{CLR Type, or a (character) type name that can be successfully parsed}
}
\value{
a list of constructor signatures
}
\description{
List the public constructors of a CLR Type
}
\examples{
\dontrun{
library(rClr)
testClassName <- "Rclr.TestObject";
clrGetConstructors(testClassName)
}
}

