% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{clrGetExtPtr}
\alias{clrGetExtPtr}
\title{Gets the external pointer CLR object.}
\usage{
clrGetExtPtr(clrObject)
}
\arguments{
\item{clrObject}{a S4 object of class clrobj}
}
\value{
the external pointer to the CLR object
}
\description{
Gets the external pointer CLR object. End user are unlikely to need this.
}

