% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{clrGetMethods}
\alias{clrGetMethods}
\title{List the instance methods of a CLR object}
\usage{
clrGetMethods(clrobj, contains = "")
}
\arguments{
\item{clrobj}{CLR object}

\item{contains}{a string that the methods names returned must contain}
}
\value{
a list of names of the methods of the CLR object
}
\description{
List the instance methods of a CLR object
}
\examples{
\dontrun{
library(rClr)
testClassName <- "Rclr.TestObject";
testObj <- clrNew(testClassName)
clrGetMethods(testObj)
clrGetMethods(testObj, 'Get')
}
}

