% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clrGetNativeLibName}
\alias{clrGetNativeLibName}
\title{Gets the name of the native library currently loaded.}
\usage{
clrGetNativeLibName()
}
\value{
the name of the native library currently loaded: rClrMs or rClrMono
}
\description{
Gets the name of the native library currently loaded. Used only for unit tests.
}

