\name{clrGetType}
\alias{clrGetType}
\title{Gets the type of a CLR object given its type name}
\usage{
  clrGetType(typename)
}
\arguments{
  \item{typename}{a character vector of length one. It can
  be the full file name of the assembly to load, or a fully
  qualified assembly name, or as a last resort a partial
  name.}
}
\value{
  the CLR Type.
}
\description{
  Gets the type of a CLR object given its type name
}

