% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{clrIs}
\alias{clrIs}
\title{Check whether an object is of a certain type}
\usage{
clrIs(obj, type)
}
\arguments{
\item{obj}{an object}

\item{type}{the object type to check for. It can be a character, of a object of CLR type System.RuntimeType}
}
\value{
TRUE or FALSE
}
\description{
Check whether an object is of a certain type. This function is meant to match the behavior of the 'is' keyword in C#.
}
\examples{
\dontrun{
library(rClr)
testClassName <- "Rclr.TestObject";
(testObj <- clrNew(testClassName))
clrIs(testObj, testClassName)
clrIs(testObj, 'System.Object')
clrIs(testObj, 'System.Double')
(testObj <- clrNew('Rclr.TestMethodBinding'))
# Test for interface definitions
clrIs(testObj, 'Rclr.ITestMethodBindings')
clrIs(testObj, clrGetType('Rclr.ITestMethodBindings'))
clrIs(testObj, clrGetType('Rclr.TestMethodBinding'))
clrIs(testObj, clrGetType('System.Reflection.Assembly'))
}
}

