% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{clrSetEnumProperty}
\alias{clrSetEnumProperty}
\title{Sets the value of an enum field or property of an object or class}
\usage{
clrSetEnumProperty(objOrType, name, enumtype, enumval)
}
\arguments{
\item{objOrType}{a CLR object, or type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}

\item{name}{the name of a field/property of the object}

\item{enumtype}{type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}

\item{enumval}{the value to set the field with}
}
\description{
Sets the value of an enum field or property of an object or class
}

