% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{clrShutdown}
\alias{clrShutdown}
\title{Shuts down the current runtime.}
\usage{
clrShutdown()
}
\value{
nothing is returned by this function
}
\description{
Shuts down the current runtime.
}

