% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{clrToString}
\alias{clrToString}
\title{Calls the ToString method of an object}
\usage{
clrToString(x)
}
\arguments{
\item{x}{any R object, which is converted to a CLR object on which to call ToString}
}
\value{
the string representation of the object in the CLR
}
\description{
Calls the ToString method of an object as represented in the CLR.
This function is here to help quickly test object equivalence from the R interpreter, for instance on the tricky topic of date-time conversions
}
\examples{
\dontrun{
library(rClr)
dt <- as.POSIXct('2001-01-01 02:03:04', tz='UTC')
clrToString(dt)
}
}

