% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{clrTypeNameExtPtr}
\alias{clrTypeNameExtPtr}
\title{Gets the type name of an object}
\usage{
clrTypeNameExtPtr(extPtr)
}
\arguments{
\item{extPtr}{external pointer to a CLR object (not a cobjRef S4 object)}
}
\value{
a character string, the type name
}
\description{
Gets the type name of an object, given the SEXP external pointer to this CLR object.
}
\examples{
\dontrun{
library(rClr)
testClassName <- "Rclr.TestObject";
testObj <- clrNew(testClassName)
clrTypeNameExtPtr(clrGetExtPtr(testObj))
}
}

