% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{clrVT}
\alias{clrVT}
\title{Get the COM variant type of a CLR object}
\usage{
clrVT(objOrType, methodName, ...)
}
\arguments{
\item{objOrType}{a CLR object, or type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}

\item{methodName}{the name of the method called on the object or class specified with objOrType}

\item{...}{one or more arguments to pass to the function call}
}
\value{
A string
}
\description{
Get the COM variant type of a CLR object, e.g. "VT_ARRAY | VT_I8". This function only works when run on the Microsoft implementation of the CLR.
This function is useful for advanced diagnosis; most users should never have to use it.
}
\examples{
\dontrun{
library(rClr)
cTypename <- "Rclr.TestCases"
#         public static bool IsTrue(bool arg)
clrVT(cTypename, 'IsTrue', TRUE)
clrVT('System.Convert', 'ToInt64', 123L)
clrVT('System.Convert', 'ToUInt64', 123L)
}
}

