\name{.onAttach}
\alias{.onAttach}
\title{rClr .onAttach}
\usage{
  .onAttach(libname = "~/R", pkgname = "rClr")
}
\arguments{
  \item{libname}{the path to the library from which the
  package is loaded}

  \item{pkgname}{the name of the package.}
}
\description{
  Print startup messages from package onLoad
}
\details{
  Print startup messages from package onLoad (prevents a
  'NOTE' on package check)
}

