% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.R
\name{dotOnLoad}
\alias{.onLoad}
\alias{dotOnLoad}
\title{rClr .onLoad}
\usage{
.onLoad(libname = "~/R", pkgname = "rClr")
}
\arguments{
\item{libname}{the path to the library from which the package is loaded}

\item{pkgname}{the name of the package.}
}
\description{
Function called when loading the rClr package with 'library'.
}
\details{
The function looks by default for the rClr native library for the Mono runtime.
If the platform is Linux, this is the only option. If the platform is Windows, using the
Microsoft .NET runtime is an option. If the rClr native library for MS.NET is detected,
the Microsoft .NET runtime is loaded in preference to Mono.
}

