% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getClrVersionString}
\alias{getClrVersionString}
\title{Gets the version of the common language runtime in use}
\usage{
getClrVersionString()
}
\value{
the version of the common language runtime in use
}
\description{
Gets the version of the common language runtime in use.
}

