% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{getCurrentConvertedObject}
\alias{getCurrentConvertedObject}
\title{System function to get a direct access to an object}
\usage{
getCurrentConvertedObject()
}
\value{
a CLR object
}
\description{
This function needs to be exported, but is highly unlikely to be of any use to an end user, even an advanced one.
This is indirectly needed to unlock the benefits of using R.NET convert data structures between R and .NET.
Using this function is a critical part of solving the rather complicated issue rClr #33.
}

