% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{getSexpType}
\alias{getSexpType}
\title{Get the type code for a SEXP}
\usage{
getSexpType(sexp)
}
\arguments{
\item{sexp}{an R object}
}
\value{
the type code, an integer, as defined in Rinternals.h
}
\description{
Get the type code for a SEXP, as returned by the TYPEOF macro
}

