% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{peekClrArgs}
\alias{peekClrArgs}
\title{Peek into the types of CLR objects arguments are converted to by rClr}
\usage{
peekClrArgs(...)
}
\arguments{
\item{...}{method arguments passed to .External}
}
\value{
a character message with type information about each argument.
}
\description{
Advanced use only, to diagnose unexpected conditions in CLR method calls. Most users would not ever need it.
}
\examples{
\dontrun{
library(rClr)
peekClrArgs("a", numeric(0))
}
}

