\docType{data}
\name{reflectionHelperTypeName}
\alias{reflectionHelperTypeName}
\title{Full type name of the reflection helper the interop code written in C#}
\format{chr "Rclr.ReflectionHelper"}
\usage{
  reflectionHelperTypeName
}
\description{
  Full type name of the reflection helper the interop code
  written in C#
}
\keyword{datasets}

