% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{setConvertAdvancedTypes}
\alias{setConvertAdvancedTypes}
\title{Turn on/off the conversion of advanced data types with R.NET}
\usage{
setConvertAdvancedTypes(enable = TRUE)
}
\arguments{
\item{enable}{if true enable, otherwise disable}
}
\description{
Turn on/off the conversion of advanced data types with R.NET. This will turn off the conversion of classes such as dictionaries into R lists,
as these are not bidirectional and you may want to see and manipulate external pointers to dictionaries in some circumstances.
}
\examples{
\dontrun{
library(rClr)
cTypename <- "Rclr.TestCases"
clrCallStatic(cTypename, "CreateStringDictionary")
setConvertAdvancedTypes(FALSE)
clrCallStatic(cTypename, "CreateStringDictionary")
}
}

