% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{clrCobj}
\alias{clrCobj}
\title{Create a reference object wrapper around a CLR object}
\usage{
clrCobj(obj, envClassWhere = .GlobalEnv)
}
\arguments{
\item{obj}{an object of S4 class clrObj}

\item{envClassWhere}{environment where the new generator is created.}
}
\value{
the reference object.
}
\description{
(EXPERIMENTAL) Create a reference object wrapper around a CLR object
}

