% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{clrGetStaticMembers}
\alias{clrGetStaticMembers}
\title{Gets the static members for a type}
\usage{
clrGetStaticMembers(objOrType)
}
\arguments{
\item{objOrType}{a CLR object, or type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}
}
\description{
Gets the static members for a type
}
\examples{
\dontrun{
library(rClr)
cTypename <- "Rclr.TestCases"
clrGetStaticMembers(cTypename)
testClassName <- "Rclr.TestObject";
testObj <- clrNew(testClassName)
clrGetStaticMembers(testObj)
}
}

