% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{clrGetStaticMethods}
\alias{clrGetStaticMethods}
\title{Gets the static members for a type}
\usage{
clrGetStaticMethods(objOrType, contains = "")
}
\arguments{
\item{objOrType}{a CLR object, or type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}

\item{contains}{a string that the property names returned must contain}
}
\description{
Gets the static members for a type
}

