% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{clrGetType}
\alias{clrGetType}
\title{Gets the type of a CLR object given its type name}
\usage{
clrGetType(objOrTypename)
}
\arguments{
\item{objOrTypename}{a character vector of length one. It can be the full file name of the assembly to load, or a fully qualified assembly name, or as a last resort a partial name.}
}
\value{
the CLR Type.
}
\description{
Gets the type of a CLR object given its type name
}

