\name{dotOnAttach}
\alias{dotOnAttach}
\title{rClr .onAttach}
\arguments{
  \item{libname}{the path to the library from which the
  package is loaded}

  \item{pkgname}{the name of the package.}
}
\description{
Print startup messages from package onLoad
}
\details{
Print startup messages from package onLoad (prevents a
'NOTE' on package check)
}

