% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.R
\name{dotOnAttach}
\alias{.onAttach}
\alias{dotOnAttach}
\title{rClr .onAttach}
\usage{
.onAttach(libname = "~/R", pkgname = "rClr")
}
\arguments{
\item{libname}{the path to the library from which the package is loaded}

\item{pkgname}{the name of the package.}
}
\description{
Print startup messages from package onLoad
}
\details{
Print startup messages from package onLoad (prevents a 'NOTE' on package check)
}

