% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.R
\name{getClrVersionString}
\alias{getClrVersionString}
\title{Gets the version of the common language runtime in use}
\usage{
getClrVersionString()
}
\value{
the version of the common language runtime in use
}
\description{
Gets the version of the common language runtime in use.
}

