% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{rClr-package}
\alias{rClr}
\alias{rClr-package}
\title{R accessing .NET/Mono}
\description{
Accessing the Common Language Runtime (.NET/Mono) from R
}
\details{
\tabular{ll}{
Package: \tab rClr\cr
Type: \tab Package\cr
Version: \tab 0.7-1\cr
Purpose: \tab First public release for more than a year. Many changes, most resulting from improvements to R.NET. This release is effectively Windows-only. Upcoming releases are expected to be executabe on Linux/Mono \cr
Date: \tab 2014-12-11\cr
License: \tab LGPL 2.1\cr
}

rClr is a package for low-level access from R to a Common Language
Runtime (CLR). The supported CLR implementations are Microsoft '.NET'
framework on Windows and Mono on several platforms, currently Windows and
Debian Linux. rClr has been used for scientific modelling work since
2013-06. Running it on Mono is still relatively experimental, but has
significantly improved over 2014.

Kosei Abe (https://www.codeplex.com/site/users/view/kos59125) is gratefully
acknowledged as the original author of the R.NET C# library, reused by this package.
}
\author{
Jean-Michel Perraud \email{jean-michel.perraud_at_csiro.au}
}
\keyword{.NET}
\keyword{CLR}
\keyword{Mono}
\keyword{package}

