% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{rClr-package}
\alias{rClr}
\alias{rClr-package}
\title{R accessing .NET/Mono}
\description{
Accessing the Common Language Runtime (.NET/Mono) from R
}
\details{
\tabular{ll}{
Package: \tab rClr\cr
Type: \tab Package\cr
Version: \tab 0.6-2\cr
Date: \tab 2014-07-10\cr
License: \tab LGPL 2.1\cr
}

A low-level, in-process interoperability package between R
and a Common Language Runtime (.NET or Mono).
The supported CLR implementations are Microsoft '.NET' framework on Windows and Mono
on several platforms, currently Windows and Debian Linux. rClr has been used for
production work as of 2013-06-16. Running it on Mono is possible, though less mature than on MS.NET.
The approach is similar to that of rJava.

Kosei Abe (https://www.codeplex.com/site/users/view/kos59125) is gratefully
acknowledged as the author of the R.NET C# library, reused by this package.
}
\author{
Jean-Michel Perraud \email{jean-michel.perraud_at_csiro.au}
}
\keyword{.NET}
\keyword{CLR}
\keyword{Mono}
\keyword{package}

