% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rClr-exported.R
\name{setRDotNet}
\alias{setRDotNet}
\title{Turn on/off R.NET}
\usage{
setRDotNet(setit = TRUE)
}
\arguments{
\item{setit}{if true enable, otherwise disable}
}
\description{
Turn on or off the usage of the R.NET assemblies to convert CLR objects to R data structures. As of version 0.7.0, R.NET is the preferred way to convert data and is enabled by default.
}
\examples{
\dontrun{
library(rClr)
## R.NET is currently used to convert complicated CLR types to sensible R equivalents
setRDotNet()
cTypename <- "Rclr.TestCases"
clrCallStatic(cTypename, "CreateStringDictionary")
setRDotNet(FALSE)
clrCallStatic(cTypename, "CreateStringDictionary")
}
}

